added type casts to get rid of compiler warnings
Former-commit-id: 4baff4564022217460b19c7a9b252415538f352c

+++ src/postgres/interfaces/libpq/fe-protocol3.cpp
@@ -1188,7 +1188,9 @@ reportErrorPosition(PQExpBuffer msg, const char *query, int loc, int encoding)
 		 * width.
 		 */
 		scroffset = 0;
-		for (; i < msg->len; i += pg_encoding_mblen(encoding, &msg->data[i]))
+		// TODO: Peloton changes - added (int) cast
+		// to suppress compiler warning
+		for (; i < (int)msg->len; i += pg_encoding_mblen(encoding, &msg->data[i]))
 		{
 			int			w = pg_encoding_dsplen(encoding, &msg->data[i]);
